Assistant checkpoint: Update input field to handle larger numbers
Assistant generated file changes:
- client/src/components/calculators/BondsTransferCostsCalculator.tsx: Update input field to handle larger numbers

---

User prompt:

This input field seems to be restricting me to one digit and two decimals please fix it so I can accommodate millions

+++ client/src/components/calculators/BondsTransferCostsCalculator.tsx
@@ -135,9 +135,9 @@ export default function BondsTransferCostsCalculator({ onCalculate }: BondsTrans
                     <Input
                       {...field}
                       onChange={(e) => {
-                        const value = e.target.value.replace(/[^0-9.]/g, "");
+                        const value = e.target.value.replace(/[^0-9]/g, "");
                         const formattedValue = value
-                          ? formatCurrency(parseFloat(value))
+                          ? formatCurrency(parseInt(value))
                           : "";
                         field.onChange(formattedValue);
                       }}
